EL-85 XES-Parsen in Worker ausgelagert (mit Fallback auf XES-Parser-Service)
