+++ src/app/workers/xes-parser.worker.ts
@@ -1,0 +1,8 @@
+/// <reference lib="webworker" />
+import {XesParser} from "../classes/EventLog/xesParser";
+
+addEventListener('message', ({ data }) => {
+    const parser = new XesParser();
+    const result = parser.parse(data);
+    postMessage(result);
+});
