+++ tsconfig.app.json
@@ -13,5 +13,8 @@
   ],
   "include": [
     "src/**/*.d.ts"
-  ]
+  ],
+    "exclude": [
+        "src/**/*.worker.ts"
+    ]
 }
